+++ keyboards/handwired/symmetric70_proto/matrix_fast/matrix.c
@@ -230,5 +230,5 @@ bool matrix_scan(void) {
     MATRIX_DEBUG_SCAN_START();
     matrix_scan_quantum();
     MATRIX_DEBUG_SCAN_END();
-    return (uint8_t)changed;
+    return changed;
 }
